[IMP] {pos,website}_event: make event questions m2m
This commit allows to create questions that can be shared by events and also by
event templates. Therefore, if a question has to be used by several events or
event templates, there is no need to duplicate it. From now on, if a question
is modified, the change will be applied to both the events and the event types
to which it is linked.

task-4247015