[PERF] core: file_path does not require exists() call
`file_path` does not always need to check that the file exists. As its
documentation indicates, it may raise such an exception. We are keeping
its behaviour unchanged, but add the possibility to skip the check. This
is possible for absolute paths and for relative paths to imported addons
(where there is only 1 path to check).

Then `file_open` does not have to check the existence of the file,
because it already uses `open` which will raise if the file does not
exist.