[IMP] crm: rearrange recurring revenues and currencies
Previously, on the CRM lead view, stages would display the
aggregated projected revenue in the column header, but
wouldn't display its currency.

Additionally, if there were aggregated recurring revenues,
the aggregated revenue would "hang" awkwardly in the middle
of the column header.

This commit makes it so aggregated revenues and recurring revenues
are packed at the end of the column header. Additionnally,
aggregated and recurring revenues display their currency.

task-4684660