[ADD] website: add technical/specification page
Before this commit:
Users may struggle to edit specific pages used by odoo website. At the moment,
they have to go to that specific page to change it, but there's no way to list
them all and this can lead to a default layout on important pages because
users don't know how to access them.

After this commit:
we can list out all that pages that are editable so we can easily find that
static pages so that user can easily find in this list and edit that pages.

task: 4712084