[IMP] website: adds color customization options in description of map snippet
In the default behaviour of the map description, there was no option to
customize the background or text color of the attached description.

This commit introduces customization options for both text and
background colors in the map snippet description. These options will be
available when the description option is enabled in the map snippet.

task-4604018