[FIX] spreadsheet: ensure that fixedPeriod is correct
Some migrations (for example `antepenultimate_year`) introduced a
fixedPeriod with a yearOffset different from 0, which was not
expected by the global filters side panel. (It worked fine by chance in
the model). This commit ensures that the fixedPeriod is always correct
when the migration is applied.

closes odoo/odoo#212884

Task: 4848233
Related: odoo/upgrade#7789
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>