[FIX] website: prevent error when multiple views has the same key
When multiple `ir.ui.view` records exist with the same `key`
(e.g., due to duplication), It can lead to a singleton error.

**Steps to reproduce:**
1. Install the `website` module.
2. Go to `Settings > Technical > User Interface > Views`.
3. Find and duplicate the **Home** view for **My Website** (same key).
4. Open the website home page in **Editor** mode.

**Error:**
`ValueError - Expected singleton: ir.ui.view(2776, 2774)`

**Cause:**
In `viewref()`, it uses `filter_duplicate()` to filter for the most suitable
view, but it may return multiple views if more than one match the criteria. - [1]

[1] - https://github.com/odoo/odoo/blob/edfa37271a0015a0d4acb17e6985a87e707e5f33/addons/website/models/website.py#L1228-L1230

**Fix:**
This commit ensures that the result is a recordset with at most one view and
preventing singleton-related errors.

Sentry - 6223988092