[IMP] website_sale: submit form through main website checkout button
Use the main website checkout button to submit the form rendered on the side template.
extra_form.js submits the form that is selected on the extensions.
Ex: in extra_info.js, the selector links which form is gonna be submitted.

Remove _cta_classes and _form_send_navigation parameters as not used anymore.